Camera and matrix update
- Replaced previous gauss-jordan method with a version made following a
tutorial (Matrix.gj2). Source in comments. I haven't yet decided if I'll
keep it or use it to correct my own version, Matrix.gaussJordan which
doesn't sort the rows. 

- Made some progress on camera rotation. The camera can now be rotated
around the x-axis by pressing up and down. A rounding error somewhere in
the gj2 method seems to be causing a noticable bug with positioning of
the bodies.