Fixed the bug that made planets jump in certain directions.
The problem occured when trying to project the planets movements into
the 2D space. When the focal length was realistic the distances would
become too small to be represented accurately with a Double. I fixed it
by making the focal length 2 meters instead and I'll probably keep it
that way instead of using BigDecimal values.