Filemanager and gui-update
Added a filemanager, a fileformat and a deserializer. I decided that it
was easier without json. Removed the example filesystem from the program
code and moved it to its own file. The program can now be tested by
running SimulatorApp.scala and controlling the rotation matrices by
using the arrows, or the right side panel.